pyproject.toml
[project]
name = "planaria-one"
version = "0.1.0"
description = "PLANARIA_ONE single-file agent engine and evaluator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.32.0",
  "tqdm>=4.66.0",
  "duckduckgo-search>=6.0.0",
  "SRTrain>=1.0.0",
]

[tool.uv]
package = false
